Refactor audio state to be in App.
