+++ src/components/Hero/Hero.scss
@@ -10,7 +10,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    background-color: rgba(0,0,0,0.3);
+    background-color:rgba(0,0,0,0.3);
     height: 100%;
     width: 100%;
     padding: 1rem;
