+++ src/components/RadioPlayer/RadioControlPanel.jsx
@@ -27,9 +27,9 @@ export default function RadioControlPanel(props) {
         <div className="RadioControlPanel-infoContainer RadioControlPanel-infoContainer--Details">
           <div className="RadioControlPanel-contactWrapper">
             <h2>Turun Wappuradio</h2>
-            {/* <div>
+            <div>
               <span>Taajuudella</span> <b>93,8 MHz</b>
-            </div> */}
+            </div>
             <div>
               <span> Studio</span> <b>023 619 0420</b>
             </div>
