+++ src/components/RadioPlayer/RadioControlPanel.scss
@@ -117,6 +117,7 @@
       align-self: flex-end;
       width: fit-content;
       margin-left: auto;
+      color: $colorWhite;
 
       > * {
         margin: 0.125rem;
