+++ src/components/RadioPlayer/RadioPlayer.scss
@@ -38,7 +38,6 @@
 
 .RadioPlayer-contact {
   padding: 2rem;
-  margin: 2rem;
 
   > p, h1 {
     font-size: 1rem;
@@ -76,6 +75,7 @@
   align-items: center;
   width: fit-content;
   font-size: 1rem;
+  font-weight: 400;
   text-decoration: none;
 
   > svg {
